TST: stats: differential_entropy: move tests to more appropriate file; ensure that they pass after refactoring
